Scraping info about games from Steam website.
This code scrapes info about popular new releases from the Steam website. The info contains the game title, price, date of release and compatible platforms. I used lxml for this.
This is just a basic automation, I plan on using it later for a proper project with an interactive front end.